+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 
 [//]: # (Removed, Added, Changed, Fixed, Updated)
 
+
+## [0.12.0] - 2024-06-29
+
 ### Fixed
 - fix the broken test-main dependencies in the intermediate KMP source sets.
 
@@ -341,6 +344,7 @@ _Stabilization release._
 
 ## Notes
 
+[0.12.0]: https://github.com/fluxo-kt/fluxo-kmp-conf/releases/tag/v0.12.0
 [0.11.0]: https://github.com/fluxo-kt/fluxo-kmp-conf/releases/tag/v0.11.0
 [0.10.2]: https://github.com/fluxo-kt/fluxo-kmp-conf/releases/tag/v0.10.2
 [0.10.1]: https://github.com/fluxo-kt/fluxo-kmp-conf/releases/tag/v0.10.1